removed UtilsExtented.h and moved its only function to Utils.h
